Complete combine game blog post.
