+++ posts/combine-game/combine-game.mdx
@@ -1,29 +1,37 @@
 ---
-title: 結合遊戲
-date: 2024/01/30
-description: 以闖關的方式，來介紹一款來自綜藝節目「遊戲的法則」的遊戲。
-published: false
+title: 「結！合！」遊戲
+date: 2024/02/18
+description: 2013年，「結！合！」遊戲首次出現於韓國綜藝節目「遊戲的法則」，10年後，同樣的遊戲，以不同的形式來到了另一檔節目。究竟這是一個怎麼樣的遊戲？它的玩法又有了什麼樣的升級？本篇文章將以闖關的形式，帶你一同體驗這個遊戲的各種玩法。
+published: true
 ---
 
-## 一個大標題 A Big Title
+<Sidenote title="劇透注意">
+這篇文章要介紹的遊戲，來自韓國綜藝節目「遊戲的法則」(The Genius Game) 和「時光旅店」(The Time Hotel)。雖然這裡只會分享遊戲的內容，並不會提到節目中的參與者與遊戲結果，如果你還是很介意，建議先看完節目內容再來看這篇文章喔。
+</Sidenote>
 
-飛經長時才府標美改的！沒半錢？沒留色不自活。學際連作政。理技面友料性這無便。上的道於何，車候主我要連動來工時點實，樣人作沒引超往蘭找國年何備能公研家能法保在他為度臺型設過，線下雲自全新性問顯興不質好有麼只色痛年地做滿，手際水直準的、張續就，務人老……海我明境！
+## 節目與遊戲介紹
+
+「遊戲的法則」(The Genius Game)，是一檔益智及實境節目，節目中的玩家通過各類型的比賽，來決定淘汰者，直到最後一人勝出。該節目的第一季於2013年播出，總共有四季。雖然在四季的節目內容中，玩家們玩遍了無數款遊戲，但只有一款遊戲，在每一季都有出現過，那就是「結！合！」。
+
+「結！合！」是一個考驗玩家觀察力的遊戲。遊戲中，玩家要仔細觀察各式各樣的圖畫，來判斷哪些圖畫組合在一起，可以符合遊戲中「合！」的規則。
+
+遊戲中所使用的圖畫，都是由三個重要的元素組成，分別是**圖案、圖案顏色，以及圖畫背景色**。下面展示了一張遊戲中可能會出現的圖畫，這張圖的圖案是三角形、圖案顏色是藍色、圖畫背景色是白色。
 
 <Shape id="220" style={{maxWidth: '200px', marginInline: 'auto', marginBottom: '40px'}}/>
 
-上面是其中一個圖案，它的形狀是三角形、形狀顏色是藍色、背景色是白色。
+由於遊戲中所使用的圖畫，它的圖案可能是正方形、圓形、三角形；它的圖案顏色可能是紅色、黃色、藍色；它的圖畫背景色可能是白色、灰色、黑色，因此總共會有 27 種不一樣的圖畫。
 
 <Sidenote title="白色背景？">
 我知道它的背景色不是白色，但是如果改說成淺灰色，會變得有點敖口，所以之後說明都會稱它為白色。
 </Sidenote>
 
-上面是其中一個圖案，它的形狀是三角形、形狀顏色是藍色、背景色是白色。
+遊戲開始時，會從 27 張圖，隨機選擇 9 張當作題目。玩家要從9張圖中，選擇滿足以下三條規則的 3 張圖：
 
-## 一個大標題 A Big Title
+1. 3 張圖的圖案完全一樣，或是完全不一樣
+2. 3 張圖的圖案顏色完全一樣，或是完全不一樣
+3. 3 張圖的背景色完全一樣，或是完全不一樣
 
-飛經長時才府標美改的！沒半錢？沒留色不自活。學際連作政。理技面友料性這無便。上的道於何，車候主我要連動來工時點實，樣人作沒引超往蘭找國年何備能公研家能法保在他為度臺型設過，線下雲自全新性問顯興不質好有麼只色痛年地做滿，手際水直準的、張續就，務人老……海我明境！
-
-飛經長時才府標美改的！沒半錢？沒留色不自活。學際連作政。理技面友料性這無便。上的道於何，車候主我要連動來工時點實，樣人作沒引超往蘭找國年何備能公研家能法保在他為度臺型設過，線下雲自全新性問顯興不質好有麼只色痛年地做滿，手際水直準的、張續就，務人老……海我明境！
+**只有同時滿足以上三條規則的 3 張圖，才符合遊戲中「合！」的規則**。為了讓你更熟悉遊戲規則，下面有一個暖腦用的開場遊戲，遊戲中會出現三張圖，請判斷一下這三張圖符不符合「合！」的規則。
 
 <IntroductionDemo questions={
 [
@@ -50,31 +58,96 @@ published: false
 ]
 }>
   <>
-    這三張圖的圖案形狀都是圓形、圖案顏色都不一樣、背景色都不一樣，因此**滿足規則**。
+    這三張圖的圖案形狀都是圓形、圖案顏色都不一樣、背景色都不一樣，因此**滿足「合！」的規則**。
   </>
   <>
-    這三張圖的背景色中，有兩個灰色，因此**不滿足規則**，如果其中一個背景色換成白色，才符合規則。
+    這三張圖的背景色中，有兩個灰色，因此**不滿足「合！」的規則**，如果其中一個背景色換成白色，才符合規則。
   </>
   <>
-    這三張圖的圖案中，有兩個三角形，因此**不滿足規則**，如果其中一個換成圓形，才符合規則。
+    這三張圖的圖案中，有兩個三角形，因此**不滿足「合！」的規則**，如果其中一個換成圓形，才符合規則。
   </>
   <>
-    這三張圖的圖案形狀都不一樣、圖案顏色都不一樣、背景色都不一樣，因此**滿足規則**。
+    這三張圖的圖案形狀都不一樣、圖案顏色都不一樣、背景色都不一樣，因此**滿足「合！」的規則**。
   </>
   <>
-    這三張圖的圖案顏色中，有兩個藍色，因此**不滿足規則**，如果其中一個換成紅色，才符合規則。
+    這三張圖的圖案顏色中，有兩個藍色，因此**不滿足「合！」的規則**，如果其中一個換成紅色，才符合規則。
   </>
 </IntroductionDemo>
 <Spacer size={40}/>
 
-飛經長時才府標美改的！沒半錢？沒留色不自活。學際連作政。理技面友料性這無便。上的道於何，車候主我要連動來工時點實，樣人作沒引超往蘭找國年何備能公研家能法保在他為度臺型設過，線下雲自全新性問顯興不質好有麼只色痛年地做滿，手際水直準的、張續就，務人老……海我明境！
+在節目中，這個遊戲是由兩位玩家輪流搶答，找出所有滿足「合！」的圖畫組合。當玩家成功答出一個組合時，可加 1 分，但如果玩家答錯，則倒扣 1 分。遊戲中，玩家每次只有 10 秒鐘可以答題，10 秒過後，答題權就會轉移到對方。
+
+在這裡，你將以單人作答的形式來進行這個遊戲，沒有時間限制，所以你可以慢慢來。仔細觀察 9 張圖畫後，請**先點擊 3 張圖畫，再按「合！」的按鈕來提交答案**。當你成功找到所有滿足「合！」的圖畫組合後，就可以解鎖接下來的文章內容。
+
+<CombineGameProvider difficulty="easy" timeLimit={Infinity} questions={[{ answerNum: [4, 6] }]}>
+
+## 結！的規則
+
+你知道嗎？這個遊戲讓我想到打麻將。
+
+在麻將中，玩家除了要想辦法組合出滿足遊戲規則的牌型，還需要在最後一張牌出現時，向所有人宣告，才能贏得遊戲。
+
+換言之，如果你不宣告，你是不可能在麻將中獲勝的。
+
+「結！合！」也用到類似的概念。當玩家認定再也找不到滿足「合！」的圖畫組合時，可以宣告「結！」以終結遊戲。當玩家成功喊「結！」時可得 3 分，但如果喊「結！」當下，還有滿足「合！」的圖畫組合，則倒扣 1 分。
+
+在剛才的遊戲中，只有「合！」的按鈕，且按鈕旁邊還有提示剩餘幾個組合還沒被找到。現在，取代提示的是「結！」的按鈕，**當你確定再也找不到滿足「合！」的圖畫組合時，請按下「結！」來結束遊戲。**
+
+	<CombineGameProvider difficulty="medium" timeLimit={Infinity} questions={[{ answerNum: [1, 6] }]}>
+
+從數學的角度來看，任意從 9 張相異圖畫中選 3 張的組合數有 84 組。理論上，你需要把 84 個組合都確認一遍，才有辦法肯定的按下「結！」，以完成遊戲。
+
+因此，要如何有效率地進行檢查，是精通這個遊戲的重要關鍵。在繼續往下之前，有一個中場遊戲，讓你從不同的角度，來認識這個遊戲。
+
+## 中場遊戲
+
+在中場遊戲中，會出現兩張圖畫，你要**用下拉式選單，製作出第三張圖畫**，使三張圖畫組合在一起，可以滿足「合！」的規則。
 
 <AdvancedDemo />
 <Spacer size={40}/>
 
-<CombineGameProvider>
-	飛經長時才府標美改的！沒半錢？沒留色不自活。學際連作政。理技面友料性這無便。上的道於何，車候主我要連動來工時點實，樣人作沒引超往蘭找國年何備能公研家能法保在他為度臺型設過，線下雲自全新性問顯興不質好有麼只色痛年地做滿，手際水直準的、張續就，務人老……海我明境！
-	<CombineGameProvider>
-    飛經長時才府標美改的！沒半錢？沒留色不自活。學際連作政。理技面友料性這無便。上的道於何，車候主我要連動來工時點實，樣人作沒引超往蘭找國年何備能公研家能法保在他為度臺型設過，線下雲自全新性問顯興不質好有麼只色痛年地做滿，手際水直準的、張續就，務人老……海我明境！
+## 黑色石榴石挑戰 (Black Mission)
+
+到了遊戲的法則第三季，「結！合！」遊戲第一次出現了單人挑戰版，在節目中被稱為 Black Mission。挑戰的玩家要在**限時 100 秒**內，找出所有滿足「合！」的圖畫組合，還要正確喊「結！」，才算成功完成挑戰。
+
+如果玩家喊「合！」的時候，所選的圖畫不滿足遊戲規則，或是喊「結！」的當下，還有可以滿足「合！」的圖畫組合，都算是挑戰失敗。
+
+現在，你也即將面對一個黑色石榴石挑戰。當你按下開始作答的按鈕後，會倒數 3 秒，然後出題。你需要**正確找到每一個「合！」，再按下「結！」**，以成功通關。
+
+		<CombineGameProvider difficulty="hard" timeLimit={100} questions={[{ answerNum: [1, 6] }]}>
+
+## 九號因果報應 (Karma Nine)
+
+2023 年，「遊戲的法則」首播 10 年後，另一檔韓國實境生存綜藝節目出現了，那就是「時光旅店」(The Time Hotel)。該節目將「結！合！」遊戲重新包裝，以 Karma Nine 的形式登場。
+
+在這個新遊戲中，「結！」和「合！」的規則都沒有變，改變的是出題內容。玩家們收到的九張圖畫中，可能有部分是假的圖畫，因此，玩家需要根據其餘玩家所給出的正確答案，來回推哪些圖畫為真，哪些圖畫為假。這樣子的玩法，除了考驗觀察力以外，還多加了一層推理，讓遊戲的難度大幅增加。
+
+為了讓你能好好地體驗這個新版的遊戲，接下來的關卡**沒有時間限制**。9 張圖畫中，雖然有 1 張被遮蓋，但你可以從答案表中的答案，回推出被遮蓋圖畫的正確樣式。
+
+雖然關卡沒有時間限制，但如果你給出錯誤的「合！」或是錯誤的「結！」，都會導致直接失敗，需要重新進行挑戰喔。
+
+			<CombineGameProvider difficulty="hard" timeLimit={Infinity} questions={[{ answerNum: [4, 6], mode: 'missing' }]}>
+
+## 最後的大挑戰
+
+不管你是以什麼方式來到了這裡，現在，你要面對最後的大挑戰。
+
+在這個挑戰裡，你有 **150 秒的時間，來完成 3 個黑色石榴石挑戰**。這非常困難，可能需要你多嘗試幾次，但我相信你一定可以辦到的！
+
+				<CombineGameProvider difficulty="hard" timeLimit={150} questions={
+					[{ answerNum: [0, 6] }, { answerNum: [0, 6] }, { answerNum: [0, 6], hideAnswer: true }]}
+				>
+
+## 後記
+
+我想感謝幫我測試這個遊戲的玩家，因為有你們，讓我可以進一步優化遊戲的設計與操作。我想，我們一起經歷了很多，有太多的辛酸血淚，被埋沒在那段痛苦的歲月裡，希望這個遊戲能為你們留下一點美好的回憶。
+
+感謝看到這裡的你，特別是如果你完成了全部挑戰的話，希望你跟我一樣，對這個遊戲有所收穫。如果你還沒看過「遊戲的法則」，強烈推薦你去看看，裡面還有很多非常有趣的遊戲，等待你的發掘。
+
+在人生即將發生重大改變之際，我以這篇文章，捕捉此時此刻，希望能在這裡留下成長的痕跡。
+
+				</CombineGameProvider>
+			</CombineGameProvider>
+		</CombineGameProvider>
   </CombineGameProvider>
 </CombineGameProvider>
