+++ src/components/OrderedList/OrderedList.js
@@ -4,8 +4,7 @@ import styled from 'styled-components';
 const OrderedList = styled.ol`
 	counter-reset: counts 0;
 	padding-left: 8px;
-	margin-top: 28px;
-	margin-bottom: 60px;
+	margin-bottom: 40px;
 `;
 
 export default OrderedList;
