+++ src/components/UnorderedList/UnorderedList.js
@@ -4,8 +4,7 @@ import { QUERIES } from '@/constants';
 
 const UnorderedList = styled.ul`
 	padding-left: 28px;
-	margin-top: 28px;
-	margin-bottom: 60px;
+	margin-bottom: 40px;
 
 	@media ${QUERIES.phoneAndDown} {
 		padding-left: 12px;
